Player layer integrated!
-- Description --
The demo now acknowledges a layer dedicated to the player and can be
worked on separately.

The player object has already been initialized and has a yellow box for
as a temporary placeholder.

The Vector and Sprite classes have also been updated.

-- Citations / Credits --
tile art - Kenney: http://opengameart.org/users/kenney